requirements/base.txt
pytz==2023.3  # https://github.com/stub42/pytz
python-slugify==8.0.1  # https://github.com/un33k/python-slugify
Pillow==9.5.0  # https://github.com/python-pillow/Pillow
argon2-cffi==21.3.0  # https://github.com/hynek/argon2_cffi
redis==4.5.4  # https://github.com/redis/redis-py
hiredis==2.2.2  # https://github.com/redis/hiredis-py
requests==2.29.0
celery==5.2.7
httpx==0.24.0
aiohttp==3.8.4

# Django
# ------------------------------------------------------------------------------
django==4.1.8  # pyup: < 4.2  # https://www.djangoproject.com/
django-environ==0.10.0  # https://github.com/joke2k/django-environ
django-model-utils==4.3.1  # https://github.com/jazzband/django-model-utils
django-allauth==0.54.0  # https://github.com/pennersr/django-allauth
django-crispy-forms==2.0  # https://github.com/django-crispy-forms/django-crispy-forms
crispy-bootstrap5==0.7  # https://github.com/django-crispy-forms/crispy-bootstrap5
django-redis==5.2.0  # https://github.com/jazzband/django-redis
# Django REST Framework
djangorestframework==3.14.0  # https://github.com/encode/django-rest-framework
django-cors-headers==3.14.0  # https://github.com/adamchainz/django-cors-headers
djoser==2.1.0 # https://djoser.readthedocs.io/en/latest/getting_started.html
dj-rest-auth==3.0.0 # https://dj-rest-auth.readthedocs.io/en/3.0.0/installation.html
# DRF-spectacular for api documentation
drf-spectacular==0.26.1  # https://github.com/tfranzel/drf-spectacular

#Elasticsearch
elasticsearch==7.14.0
elasticsearch-dsl==7.4.0
django-elasticsearch-dsl==7.3.0

# Cloudinary
cloudinary==1.32.0
django-cloudinary-storage==0.3.0
